login

+++ Gladiator/src/main/java/com/example/gladiator/Models/DTOs/BattleRequestDTO.java
@@ -2,6 +2,6 @@ package com.example.gladiator.Models.DTOs;
 
 public class BattleRequestDTO {
 
-    public int ownGladiator;
-    public int oponentGladiator;
+    public Long ownGladiator;
+    public Long oponentGladiator;
 }
